pyproject.toml
[tool.poetry]
name = "worldender-ai"
version = "0.1.0"
description = "Text-adventure game about world ending events"
authors = ["James Dominguez <james@keppylab.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "worldender"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.33.0"
bs4 = "^0.0.2"
instructor = "^1.3.2"
pandas = "^2.2.2"
wandb = "^0.17.1"
asyncio = "^3.4.3"
tabulate = "^0.9.0"
pymongo = "^4.7.3"

[tool.poetry.group.dev.dependencies]
mkdocs-material = "^9.5.26"
ipykernel = "^6.29.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.0.1"

[tool.poetry.group.api-server]
optional = true

[tool.poetry.group.api-server.dependencies]
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
motor = "^3.4.0"
pymongo = "^4.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
